Rework antishare
The original implementation did not work out because WC3 client blocks the
Allies tab until the game host confirms that sharing control succeeded.
Therefore, it was not possible to revert it while holding the share action.

There are now three possible values: native, kick, restrict.
Kick is simpler but may be surprising.
Restrict is very forgiving with the sharer, and proportionally abusable by griefers